Split accumulation effect into 3 different effects
