+++ js/effects/index.js
@@ -4,7 +4,9 @@ import ConvergeCircleEffect from './converge-circle';
 import WaveEffect from './wave';
 import ChangeImageEffect from './change-image';
 import FlickrImageEffect from './flickr-image';
-import Accumulation from './accumulation';
+import TrailsEffect from './trails';
+import SmoothTrailsEffect from './smooth-trails';
+import SmearEffect from './smear';
 
 const effectList = [
   HueDisplaceEffect,
@@ -13,7 +15,9 @@ const effectList = [
   WaveEffect,
   ChangeImageEffect,
   FlickrImageEffect,
-  Accumulation
+  TrailsEffect,
+  SmoothTrailsEffect,
+  SmearEffect
 ];
 const byId = {};
 for (let i = 0; i < effectList.length; i++) {
